(feat) - O3-4200 Service queues - use visit form in patient chart for create queue entry workflow
