+++ packages/esm-appointments-app/translations/en.json
@@ -7,6 +7,7 @@
   "appointmentCancelError": "Error cancelling appointment",
   "appointmentCancelled": "Appointment Cancelled",
   "appointmentCancelledSuccessfully": "Appointment cancelled successfully",
+  "appointmentCheckedIn": "Appointment Checked In",
   "appointmentColor": "Appointment color",
   "appointmentConflict": "Appointment conflict",
   "appointmentEdited": "Appointment edited",
@@ -18,6 +19,7 @@
   "appointmentEndError": "Error ending appointment",
   "appointmentFormError": "Error scheduling appointment",
   "appointmentHistory": "Appointment History",
+  "appointmentMarkedChecked": "Appointment marked as Checked In",
   "appointmentMetrics": "Appointment metrics",
   "appointmentMetricsLoadError": "",
   "appointmentNoteLabel": "Write an additional note",
@@ -158,6 +160,7 @@
   "unscheduledAppointments_lower": "unscheduled appointments",
   "upcoming": "Upcoming",
   "upcomingAppointments": "Upcoming appointments",
+  "updateError": "Error updating upcoming appointment",
   "view": "View",
   "vitals": "Vitals",
   "week": "Week",
