+++ packages/esm-service-queues-app/src/create-queue-entry/queue-fields/queue-fields.test.tsx
@@ -2,10 +2,18 @@
 import React from 'react';
 import userEvent from '@testing-library/user-event';
 import { render, screen } from '@testing-library/react';
-import { getDefaultsFromConfigSchema, useConfig, useLayoutType, useSession } from '@openmrs/esm-framework';
+import {
+  type FetchResponse,
+  getDefaultsFromConfigSchema,
+  useConfig,
+  useLayoutType,
+  useSession,
+  type Visit,
+} from '@openmrs/esm-framework';
 import { configSchema, type ConfigObject } from '../../config-schema';
-import { mockSession } from '__mocks__';
-import VisitFormQueueFields from './visit-form-queue-fields.component';
+import { mockSession, mockVisitAlice } from '__mocks__';
+import QueueFields from './queue-fields.component';
+import { postQueueEntry } from './queue-fields.resource';
 
 const mockUseConfig = jest.mocked(useConfig<ConfigObject>);
 const mockUseLayoutType = jest.mocked(useLayoutType);
@@ -32,7 +40,14 @@ jest.mock('../../hooks/useQueues', () => {
   };
 });
 
-describe('VisitFormQueueFields', () => {
+jest.mock('./queue-fields.resource', () => {
+  return {
+    postQueueEntry: jest.fn(),
+  };
+});
+const mockPostQueueEntry = jest.mocked(postQueueEntry).mockResolvedValue({} as FetchResponse);
+
+describe('QueueFields', () => {
   beforeEach(() => {
     mockUseLayoutType.mockReturnValue('small-desktop');
     mockUseSession.mockReturnValue(mockSession.data);
@@ -43,27 +58,31 @@ describe('VisitFormQueueFields', () => {
 
   it('renders the form fields and returns the set values', async () => {
     const user = userEvent.setup();
-    const setFormFields = jest.fn();
-    render(<VisitFormQueueFields setFormFields={setFormFields} />);
+    let onSubmit: (visit: Visit) => Promise<any> = null;
+    const setOnSubmit = (callback) => (onSubmit = callback);
+    render(<QueueFields setOnSubmit={setOnSubmit} />);
 
     expect(screen.getByLabelText('Select a queue location')).toBeInTheDocument();
     expect(screen.getByLabelText('Select a service')).toBeInTheDocument();
     expect(screen.getByLabelText('Sort weight')).toBeInTheDocument();
 
+    const queueUuid = 'e2ec9cf0-ec38-4d2b-af6c-59c82fa30b90';
     const serviceSelect = screen.getByLabelText('Select a service').closest('select');
-    await user.selectOptions(serviceSelect, 'e2ec9cf0-ec38-4d2b-af6c-59c82fa30b90');
+    await user.selectOptions(serviceSelect, queueUuid);
 
     expect(screen.getByText('Priority')).toBeInTheDocument();
     expect(screen.getByText('High')).toBeInTheDocument();
 
-    expect(setFormFields).toHaveBeenCalledWith({
-      queueLocation: '1',
-      service: 'e2ec9cf0-ec38-4d2b-af6c-59c82fa30b90',
-      // The status and priority are the defaults that come from the config schema.
-      // They are selected even though they are not 'allowed' for the queue.
-      status: '51ae5e4d-b72b-4912-bf31-a17efb690aeb',
-      priority: 'f4620bfa-3625-4883-bd3f-84c2cce14470',
-      sortWeight: 0,
-    });
+    await onSubmit(mockVisitAlice);
+    expect(mockPostQueueEntry).toHaveBeenCalledWith(
+      mockVisitAlice.uuid,
+      queueUuid, // queueUuid
+      mockVisitAlice.patient.uuid,
+      'f4620bfa-3625-4883-bd3f-84c2cce14470', // priority
+      '51ae5e4d-b72b-4912-bf31-a17efb690aeb', // status
+      0, // sortWeight
+      '1', // locationUuid
+      null, // visitQueueNumberAttributeUuid
+    );
   });
 });
