--- packages/esm-service-queues-app/src/patient-search/hooks/useDefaultLocation.ts
@@ -1,14 +1,0 @@
-import { type FetchResponse, openmrsFetch, useConfig } from '@openmrs/esm-framework';
-import useSWRImmutable from 'swr/immutable';
-
-export const useDefaultLoginLocation = () => {
-  const config = useConfig();
-  const apiUrl = config.defaultFacilityUrl;
-  const { data, error, isLoading } = useSWRImmutable<FetchResponse>(apiUrl, openmrsFetch);
-
-  return {
-    defaultFacility: data ? data?.data : null,
-    isLoading: isLoading,
-    error,
-  };
-};
