--- packages/esm-service-queues-app/src/patient-search/patient-scheduled-visits.scss
@@ -1,131 +1,0 @@
-@use '@carbon/layout';
-@use '@carbon/type';
-@use '@openmrs/esm-styleguide/src/vars' as *;
-
-.container {
-  background-color: $ui-background;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  height: 100%;
-}
-
-.grid {
-  padding: 0;
-}
-
-.visitTile {
-  margin: layout.$spacing-05 0;
-
-  &:global(.cds--tile--is-selected:focus) {
-    background-color: $color-blue-10;
-    color: $color-blue-10;
-    border: 1px solid $color-blue-60-2;
-    outline: none;
-  }
-
-  &:global(.cds--tile--is-selected:hover) {
-    background-color: $color-blue-10;
-    color: $color-blue-10;
-    border: 1px solid $color-blue-60-2;
-    outline: none;
-  }
-
-  &:global(.cds--tile--is-selected:hover .cds--tile__checkmark svg) {
-    fill: $color-blue-60-2;
-  }
-}
-
-.recentlyScheduledVisitsContainer {
-  margin-top: layout.$spacing-06;
-  margin-bottom: layout.$spacing-10;
-}
-
-.row {
-  margin: 0 layout.$spacing-05 layout.$spacing-05 layout.$spacing-05;
-}
-
-.backButton {
-  align-items: center;
-  display: flex;
-  justify-content: flex-start;
-  margin: layout.$spacing-03 0;
-  padding: 0;
-  @include type.type-style('body-compact-01');
-
-  button {
-    display: flex;
-
-    svg {
-      order: 1;
-      margin-right: layout.$spacing-03;
-      margin-left: 0 !important;
-    }
-
-    span {
-      order: 2;
-    }
-  }
-}
-
-.heading {
-  @include type.type-style('body-02');
-  color: $text-02;
-  margin-left: 1.2rem;
-}
-
-.primaryText {
-  @include type.type-style('heading-compact-02');
-}
-
-.secondaryText {
-  @include type.type-style('body-01');
-  color: $text-02;
-}
-
-.button {
-  height: layout.$spacing-10;
-  display: flex;
-  align-content: flex-start;
-  align-items: baseline;
-  min-width: 50%;
-}
-
-.text-divider {
-  display: flex;
-  align-items: center;
-  margin: layout.$spacing-08 25%;
-  @include type.type-style('heading-03');
-  color: $text-02;
-}
-
-.text-divider::before,
-.text-divider::after {
-  content: '';
-  height: 1px;
-  background-color: $text-03;
-  flex-grow: 1;
-}
-
-.text-divider::before {
-  margin-right: layout.$spacing-05;
-}
-
-.text-divider::after {
-  margin-left: layout.$spacing-05;
-}
-
-.buttonContainer {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin-bottom: layout.$spacing-03;
-}
-
-.prioritySwitcher {
-  margin: layout.$spacing-03 0 layout.$spacing-03;
-}
-
-.emptyAppointment {
-  margin: layout.$spacing-05;
-}
