--- packages/esm-service-queues-app/src/patient-search/visit-form/visit-type-selector.scss
@@ -1,100 +1,0 @@
-@use '@carbon/layout';
-@use '@carbon/type';
-@use '@openmrs/esm-styleguide/src/vars' as *;
-
-.visitTypeOverviewWrapper {
-  margin: layout.$spacing-05 0;
-  border: 0.0625rem solid $grey-2;
-}
-
-.tablet {
-  background-color: $ui-02;
-}
-
-.desktop {
-  background-color: $ui-01;
-
-  .paginationContainer div {
-    background-color: $ui-01;
-  }
-}
-
-.visitTypeOverviewWrapper div:nth-child(3) > div:nth-child(2) {
-  position: relative;
-}
-
-.visitTypeOverviewWrapper div:nth-child(3) span * {
-  display: none;
-}
-
-.radioButtonGroup {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  margin-top: layout.$spacing-03;
-  min-height: layout.$spacing-10;
-  width: 100%;
-  @include type.type-style('body-compact-01');
-}
-
-.radioButton {
-  padding: layout.$spacing-02 layout.$spacing-05;
-  margin: layout.$spacing-03 0;
-}
-
-.content {
-  @include type.type-style('heading-compact-01');
-  color: $text-02;
-  margin-top: layout.$spacing-05;
-  margin-bottom: layout.$spacing-03;
-}
-
-.desktopHeading {
-  h4 {
-    @include type.type-style('heading-compact-02');
-    color: $text-02;
-  }
-}
-
-.tabletHeading {
-  h4 {
-    @include type.type-style('heading-03');
-    color: $text-02;
-  }
-}
-
-.desktopHeading,
-.tabletHeading {
-  text-align: left;
-  text-transform: capitalize;
-  margin-bottom: layout.$spacing-05;
-
-  h4:after {
-    content: '';
-    display: block;
-    width: layout.$spacing-07;
-    padding-top: 0.188rem;
-    border-bottom: 0.375rem solid var(--brand-03);
-  }
-}
-
-.heading:after {
-  content: '';
-  display: block;
-  width: layout.$spacing-07;
-  padding-top: 0.188rem;
-  border-bottom: 0.375rem solid var(--brand-03);
-}
-
-.tile {
-  text-align: center;
-  border: 1px solid $ui-03;
-}
-
-// Overriding styles for RTL support
-html[dir='rtl'] {
-  .desktopHeading,
-  .tabletHeading {
-    text-align: right;
-  }
-}
