+++ packages/esm-service-queues-app/src/routes.json
@@ -127,9 +127,9 @@
       "type": "service-queues"
     },
     {
-      "name": "service-queues-patient-search",
-      "title": "searchPatient",
-      "component": "patientSearchWorkspace",
+      "name": "create-queue-entry-workspace",
+      "title": "addPatientToQueue",
+      "component": "createQueueEntryWorkspace",
       "type": "service-queues"
     }
   ]
