+++ packages/esm-service-queues-app/translations/en.json
@@ -1,6 +1,7 @@
 {
   "actions": "Actions",
   "addAProviderQueueRoom": "Add a provider queue room?",
+  "addedPatientToQueue": "Added patient to queue",
   "addNewQueueService": "Add New Queue Service",
   "addNewQueueServiceRoom": "Add new queue service room",
   "addNewService": "Add new service",
@@ -17,10 +18,8 @@
   "alistOfClients": "A list of clients waiting for ",
   "all": "All",
   "and": "And",
-  "anotherVisitType": "Start another visit type",
   "applyFilters": "Apply filters",
   "averageWaitTime": "Average wait time today",
-  "backToScheduledVisits": "Back To Scheduled Visits",
   "backToSearchResults": "Back to search results",
   "between": "Between",
   "bmi": "Bmi",
@@ -45,7 +44,6 @@
   "currentVisit": "Current visit",
   "date": "Date",
   "date&Time": "Date & time",
-  "dateAndTimeOfVisit": "Date and time of visit",
   "delete": "Delete",
   "deleteQueueEntry": "Delete queue entry",
   "discard": "Discard",
@@ -63,18 +61,14 @@
   "errorAddingQueue": "Error adding queue",
   "errorAddingQueueRoom": "Error adding queue room",
   "errorClearingQueues": "Error clearing queues",
-  "errorFetchingAppointments": "Error fetching appointments",
+  "errorFetchingVisit": "Error fetching patient visit",
   "errorLoadingQueueEntries": "Error loading queue entries",
   "errorPostingToScreen": "Error posting to screen",
-  "facility": "Facility",
-  "failedToLoadRecommendedVisitTypes": "Failed to load recommended visit types",
   "femaleLabelText": "Female",
   "filter": "Filter",
   "filterByService": "Filter by service:",
   "filterByStatus": "Filter by status:",
   "filterTable": "Filter table",
-  "futureScheduledVisits_one": "{{count}} visit scheduled for dates in the future",
-  "futureScheduledVisits_other": "{{count}} visits scheduled for dates in the future",
   "gender": "Gender",
   "heartRate": "Heart rate",
   "height": "Height",
@@ -95,13 +89,11 @@
   "missingQueueRoomName": "Missing queue room name",
   "missingQueueRoomService": "Missing queue room service",
   "missingService": "Missing service",
-  "missingVisitType": "Missing visit type",
   "modifyDefaultValue": "Modify default value",
   "movePatientToNextService": "Move patient to the next service?",
   "moveToNextService": "Move to next service",
   "name": "Name",
   "nextPage": "Next page",
-  "noAppointmentsFound": "No appointments found",
   "noColumnsDefined": "No table columns defined. Check Configuration",
   "noEncountersFound": "No encounters found",
   "noLastEncounter": "There is no last encounter to display for this patient",
@@ -120,16 +112,12 @@
   "noStatusConfigured": "No status configured",
   "notableConfig": "No table configuration",
   "notes": "Notes",
-  "noVisitTypesMatchingSearch": "There are no visit types matching this search text",
   "noVitalsFound": "No vitals found",
   "onTime": "On time",
   "orderDurationAndUnit": "for {{duration}} {{durationUnit}}",
   "orderIndefiniteDuration": "Indefinite duration",
-  "orInProperFormat": "Or",
   "patientAge": "Age",
-  "patientAlreadyInQueue": "Patient is already in the queue",
   "patientAttendingService": "Patient attending service",
-  "patientHasActiveVisit": "The patient already has an active visit",
   "patientInfo": "{{name}}{{sexInfo}}{{ageInfo}}",
   "patientList": "Patient list",
   "patientName": "Patient name",
@@ -147,7 +135,6 @@
   "priority": "Priority",
   "priorityComment": "Priority comment",
   "priorityIsRequired": "Priority is required",
-  "program": "Program",
   "provider": "Provider",
   "quantity": "Quantity",
   "queueAddedSuccessfully": "Queue added successfully",
@@ -181,9 +168,6 @@
   "queueScreen": "Queue screen",
   "queueService": "Queue service",
   "queueStatus": "Queue status",
-  "recentScheduledVisits_one": "{{count}} visit scheduled for +/- 7 days",
-  "recentScheduledVisits_other": "{{count}} visits scheduled for +/- 7 days",
-  "recommended": "Recommended",
   "refills": "Refills",
   "removeFromQueueAndEndVisit": "Remove patient from queue and end active visit",
   "removePatient": "Remove patient",
@@ -197,22 +181,15 @@
   "save": "Save",
   "scheduledAppointmentsList": "Scheduled appointments patient list",
   "scheduledToday": "Scheduled for today",
-  "search": "Search",
-  "searchForAVisitType": "Search for a visit type",
-  "searchPatient": "Search Patient",
   "searchThisList": "Search this list",
   "selectALocation": "Select a location",
   "selectAVisitType": "Select visit type",
-  "selectFacility": "Select a facility",
-  "selectOption": "Select an option",
-  "selectProgramType": "Select program type",
   "selectQueue": "Select a queue",
   "selectQueueLocation": "Select a queue location",
   "selectQueueService": "Select a queue service",
   "selectRoom": "Select a room",
   "selectService": "Select a service",
   "selectServiceType": "Select a service type",
-  "selectVisitType": "Please select a Visit Type",
   "serve": "Serve",
   "servePatient": "Serve patient",
   "service": "Service",
@@ -222,10 +199,6 @@
   "sortWeight": "Sort weight",
   "sp02": "Sp02",
   "startAgeRangeInvalid": "Start age range is not valid",
-  "startAVisit": "Start a visit",
-  "startVisit": "Start visit",
-  "startVisitError": "Error starting visit",
-  "startVisitQueueSuccessfully": "Patient has been added to active visits list and queue.",
   "status": "Status",
   "statusIsRequired": "Status is required",
   "submitting": "Submitting...",
